Updated layout component
